Batch transactions and other speedups
This PR batches the transactions necessary to create an user, which makes the user
creation much faster. It puts all instructions in one ClientTransaction and then
sends this to the conodes.

It also removes a lot of double-loading for the contacts. On startup, it will only
load the credentialInstance of each contact, the rest has to be loaded when
needed.