Make demonstrator nicer and add user-configuration
Some improvements from the UI of the demonstrator, to have a nicer user experience:
- correctly go to the c4dt-subuser config
- removed some buttons added admin/c4dt links

New user-configuration for the admin, that allows to directly set the darc
the user will be in, also adding this darc to the user.

Only give 500 credits to new users, so they can only create 5 devices,
but not new users.