Check cas-login and remove gData
This PR got bigger than it should be. It started with actually checking whether
the user exists, and then checking if the user has the right to log in. Doing
so I realized that the gData.load was in too many places, and replaced it with
UserData-service.

Now the whole load/login procedure is done in the app.component at the beginning,
and all other components (with the exception of newuser/ and register/) can
suppose that UserData holds a valid user.