Checkpoint with 'create BC' working
At this point, the conode is being created correctly,
the README explains how to use virtual interfaces to talk to
it, the logs are working right, and you can create a new
ledger and later use the info stored on your device to
reconnect later.

This version currently works to the same extent on iOS
and Android.