Checkpoint: sending coins works
Sending coins to an unregistered contact from either the
home page or from the contact list works. The next problem
is that the unregistered contact cannot update to recognize that
it is now a registered contact.

QR codes are sometimes detected as UPC codes, and that causes
a silent failure. Try to make that better.