Merge with latest dynacred and cothority
This PR follows the latest development in the dynacred and the cothority
library. Some methods had to be moved to different classes, others are
more of a cleanup of classes that had never been used before.

While cleaning up, the loading-logic has been changed. Now the
app-root holds a frame that points either to the loading and
setup pages, or that points to the tabs. This avoids problems with
differences in iOS and Android on how they handle
"visibility='hidden/collapse'" and makes for a nicer startup.
Now all loading code is done in the "loading-page", and the
"main-page" only handles the tabs.

For more information, see the CHANGELOG.