Fix a problem that blocked dev work
At least on iOS, the write in Data.ts save()
was almost always failing due to the parallel send
sending the last tx in, and then us getting a signer count
that was one too low, because the last tx was not waited
for. Always talking to the same server (parallel = 1)
works around this for the moment.