Don't allow recoveries to sign
Previously a recovery-darc has been added to both the _sign and the evolve
rule of the signer-darc. Which meant that a recovery darc was also evaluated
for siging, which is confusing.

Even though a recovery signature can theoretically recover the ID and then
use it to sign, it should not be the default behaviour.

This PR makes it so that recovery-darcs are only added to the _evolve rule
of the signer-darc, and also adds a button to change the behaviour of
the recovery-entry in the UI.