Add baseurl configuration
Now that there will be two installations - one on demo.c4dt.org to recover the accounts,
and one on login.c4dt.org - there needs to be a way to force the creation of the links
to the new installation.

This PR adds a field 'BaseURL' to the config.toml that will be used to create the links for
new users and recovery accounts.